ELT_DataScience_AWSRDS/requirements.txt
boto3
pandas
pymysql
python-dotenv
